fixed major bug in Synch
Before we stopped search upon seeing a pair of *nodes* that had already
been visited. This approach stops the search too early and misses some path
pairs.

Instead, we track the path pairs already seen and only stop the search when
we encounter previously discovered path pairs, or a node has no successors
(i.e. is a CFG exit).

This fix actually simplifies the implementation slightly because we were
already tracking the set of seen paths. Now we just use that set and avoid
tracking the set of seen node pairs.